Update file splits for `no1` `e_red_door`
