+++ config/splat.us.stno1.yaml
@@ -55,6 +55,7 @@ segments:
       - [0x281C, data]
       - [0x34670, .rodata, e_mouse]
       - [0x34670, .rodata, e_red_door]
+      - [0x34688, .rodata, unk_35E20]
       - [0x347BC, .rodata, unk_3CB34]
       - [0x34810, .rodata, e_wall_chicken_secret]
       - [0x3486C, .rodata, unk_3FA34]
@@ -75,6 +76,8 @@ segments:
       - [0x34CBC, .rodata, prim_helpers]
       - [0x34CC4, c, e_mouse]
       - [0x34E6C, c, e_red_door]
+      - [0x35C64, c, st_debug]
+      - [0x35E20, c, unk_35E20]
       - [0x3CB34, c, unk_3CB34]
       - [0x3E880, c, e_wall_chicken_secret]
       - [0x3FA34, c, unk_3FA34]
