+++ src/st/no1/st_debug.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/no1/nonmatchings/st_debug", EntityUnkId12);
